Ajout routes CREATE/PUT/DELETE SUR Comment + corrections routes Article + passage des id de SMALLINT vers MEDIUMINT sauf pour la table Users

+++ back/database/models/Articles.js
@@ -3,7 +3,7 @@ module.exports = function(sequelize, DataTypes) {
   return sequelize.define('Articles', {
     id: {
       autoIncrement: true,
-      type: DataTypes.SMALLINT.UNSIGNED,
+      type: DataTypes.MEDIUMINT.UNSIGNED,
       allowNull: false,
       primaryKey: true
     },
